requirements.txt
pillow
discord.py
pytesseract